TestHistory component and userGetTraderEvents hook
